First implementation of the custom messagebox
